refactor(runner): split main loop tests by behavior
