+++ crates/runner/src/cmd/start/tests/main_loop/admission.rs
@@ -1,0 +1,309 @@
+use super::super::super::*;
+use super::super::support::{
+    assert_run_exits_within, minimal_context, mock_run_config, mock_run_config_with_overrides,
+    push_job, shutdown, test_profiles, wait_budget_count, wait_cancel_token,
+    wait_cancel_token_removed, wait_discover_entered,
+};
+use std::sync::Arc;
+
+/// TOCTOU regression: soft drain can arrive after the main loop has selected
+/// a discovered candidate but before claim. The candidate is still unowned at
+/// that point, so Draining must roll back local admission and skip claim.
+#[tokio::test]
+async fn claim_after_draining_sent_skips_unclaimed_job() {
+    let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
+    let budget = Arc::clone(&config.capacity.budget);
+    let run_handle = tokio::spawn(run(config));
+
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    assert!(env.parking_gate.soft_drain());
+    env.mode_tx.send_if_modified(|v| {
+        *v = RunnerMode::Draining;
+        false
+    });
+
+    let run_id = RunId::new_v4();
+    push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
+
+    assert_run_exits_within(
+        run_handle,
+        Duration::from_secs(3),
+        "draining should skip the discovered unclaimed job and exit",
+    )
+    .await;
+
+    wait_cancel_token_removed(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
+    wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
+    let completions = env.handle.completions.lock().unwrap();
+    assert!(
+        !completions
+            .iter()
+            .any(|completion| completion.run_id == run_id),
+        "draining should not claim or complete a newly discovered job"
+    );
+}
+
+/// TOCTOU regression: a SIGTERM that iterates `cancel_tokens` *before*
+/// the main loop inserts a newly-claimed job's token would leave that
+/// job running uncancelled. The fix is a post-insert `mode_rx.borrow()`
+/// check that catches Stopping and cancels the token in that window.
+///
+/// To reproduce deterministically, we use `send_if_modified` to flip
+/// the watch value to `Stopping` **without** waking `mode_rx.changed()`
+/// — this is exactly what the racy window looks like to the main loop:
+/// its outer select! is still polling discover_fut, unaware that the
+/// value has changed. When discover yields a job, the main loop takes
+/// the claim path, inserts the token, then reads `mode_rx.borrow()`
+/// and catches the Stopping value that was silently written.
+#[tokio::test]
+async fn claim_after_stopping_sent_cancels_new_job() {
+    let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
+    let run_handle = tokio::spawn(run(config));
+
+    // Deterministic barrier: wait for run()'s main loop to have polled
+    // `discover_fut` into its await state. Only then is the Running-mode
+    // reactor `select!` provably in place, which is the precondition for the
+    // silent `send_if_modified` below to land without waking the loop.
+    // A wall-clock sleep here flakes under coverage CI — see #10146.
+    // The 2s timeout gives a clear diagnostic if the "loop parks on
+    // discover" invariant ever regresses, rather than hanging until
+    // the outer test harness kills us.
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    // Flip the watch value to Stopping without firing changed().
+    env.parking_gate.close();
+    env.mode_tx.send_if_modified(|v| {
+        *v = RunnerMode::Stopping;
+        false
+    });
+
+    let run_id = RunId::new_v4();
+    push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
+
+    // `wait_completion` is event-driven (fires on `provider.complete`), so
+    // this duration is a diagnostic cap for genuine hangs — not a budget
+    // for the run. A large cap absorbs coverage-CI slowdown of the full
+    // dispatch→executor→complete chain without flaking (see #10146).
+    let c = env
+        .handle
+        .wait_completion(run_id, Duration::from_secs(30))
+        .await;
+    assert!(
+        c.is_some(),
+        "job must report cancellation even when the handler missed the token"
+    );
+    assert_eq!(c.unwrap().error.as_deref(), Some("cancelled by user"));
+
+    // Let run() exit — fire changed() now so the main loop observes
+    // Stopping at loop top and breaks to teardown.
+    env.parking_gate.close();
+    env.mode_tx.send_modify(|v| {
+        *v = RunnerMode::Stopping;
+    });
+    env.cancel.cancel();
+    let _ = tokio::time::timeout(Duration::from_secs(5), run_handle).await;
+}
+
+// -----------------------------------------------------------------------
+// Test 4: Claim failure (409) rolls back budget
+// -----------------------------------------------------------------------
+
+#[tokio::test(start_paused = true)]
+async fn claim_failure_rolls_back_budget() {
+    // Budget for exactly 1 job (2 vcpu, 4096 MB matches the test profile).
+    let (config, env) = mock_run_config(test_profiles(), 2, 4096, 1);
+    let budget = Arc::clone(&config.capacity.budget);
+    let run_handle = tokio::spawn(run(config));
+
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    // First job: claim returns None (409 conflict)
+    let run_id_1 = RunId::new_v4();
+    push_job(&env, run_id_1, "vm0/default", None);
+
+    // Returning to discovery proves the failed claim was processed.
+    wait_discover_entered(&env, Duration::from_secs(5)).await;
+    wait_cancel_token_removed(&env.cancel_tokens, run_id_1, Duration::from_secs(5)).await;
+    assert_eq!(budget.allocated().2, 0);
+
+    // Second job: claim succeeds — budget should have been freed.
+    let run_id_2 = RunId::new_v4();
+    push_job(
+        &env,
+        run_id_2,
+        "vm0/default",
+        Some(minimal_context(run_id_2)),
+    );
+
+    let completion = env
+        .handle
+        .wait_completion(run_id_2, Duration::from_secs(5))
+        .await;
+    assert!(
+        completion.is_some(),
+        "second job should complete (budget freed after first 409)"
+    );
+
+    shutdown(&env, run_handle).await;
+}
+
+#[tokio::test(start_paused = true)]
+async fn claim_run_id_mismatch_rolls_back_local_state() {
+    // Budget for exactly 1 job, so a leaked lease would block the follow-up job.
+    let (config, env) = mock_run_config(test_profiles(), 2, 4096, 1);
+    let budget = Arc::clone(&config.capacity.budget);
+    let run_handle = tokio::spawn(run(config));
+
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    let candidate_run_id = RunId::new_v4();
+    let context_run_id = RunId::new_v4();
+    push_job(
+        &env,
+        candidate_run_id,
+        "vm0/default",
+        Some(minimal_context(context_run_id)),
+    );
+
+    wait_discover_entered(&env, Duration::from_secs(5)).await;
+    wait_cancel_token_removed(&env.cancel_tokens, candidate_run_id, Duration::from_secs(5)).await;
+    assert_eq!(budget.allocated().2, 0);
+    {
+        let completions = env.handle.completions.lock().unwrap();
+        assert!(
+            !completions
+                .iter()
+                .any(|completion| completion.run_id == candidate_run_id
+                    || completion.run_id == context_run_id),
+            "mismatched claim should not produce a completion for either run id"
+        );
+    }
+
+    let followup_run_id = RunId::new_v4();
+    push_job(
+        &env,
+        followup_run_id,
+        "vm0/default",
+        Some(minimal_context(followup_run_id)),
+    );
+
+    let completion = env
+        .handle
+        .wait_completion(followup_run_id, Duration::from_secs(5))
+        .await;
+    assert!(
+        completion.is_some(),
+        "follow-up job should complete after mismatched claim is rejected"
+    );
+
+    shutdown(&env, run_handle).await;
+}
+
+// -----------------------------------------------------------------------
+// Test 6: Unknown profile is skipped without affecting subsequent jobs
+// -----------------------------------------------------------------------
+
+#[tokio::test(start_paused = true)]
+async fn unknown_profile_skipped() {
+    let (config, env) = mock_run_config(test_profiles(), 8, 32768, 4);
+    let run_handle = tokio::spawn(run(config));
+
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    // Push a job with a profile that doesn't exist in the profiles map.
+    // The main loop should log a warning and continue without claiming.
+    let bad_id = RunId::new_v4();
+    push_job(
+        &env,
+        bad_id,
+        "vm0/nonexistent",
+        Some(minimal_context(bad_id)),
+    );
+
+    // The next discover wait proves the bad job was consumed and skipped.
+    wait_discover_entered(&env, Duration::from_secs(5)).await;
+
+    // Push a valid job — it should succeed despite the earlier bad one.
+    let good_id = RunId::new_v4();
+    push_job(&env, good_id, "vm0/default", Some(minimal_context(good_id)));
+
+    let completion = env
+        .handle
+        .wait_completion(good_id, Duration::from_secs(5))
+        .await;
+    assert!(
+        completion.is_some(),
+        "valid job should complete after unknown profile is skipped"
+    );
+
+    // The bad job should never have been claimed (no completion recorded).
+    {
+        let comps = env.handle.completions.lock().unwrap();
+        assert!(
+            !comps.iter().any(|c| c.run_id == bad_id),
+            "unknown-profile job should not produce a completion"
+        );
+    }
+
+    shutdown(&env, run_handle).await;
+}
+
+// -----------------------------------------------------------------------
+// Test 7: Duplicate discovery (same run_id) is deduplicated
+// -----------------------------------------------------------------------
+
+#[tokio::test(start_paused = true)]
+async fn duplicate_discovery_deduplicated() {
+    // Budget for 2 jobs — enough for the duplicate to pass the budget
+    // check and reach the cancel_tokens dedup logic.
+    let gate = Arc::new(tokio::sync::Notify::new());
+    let overrides = Arc::new(sandbox_mock::MockSandboxOverrides::with_wait_process_gate(
+        Arc::clone(&gate),
+    ));
+    let (config, env) = mock_run_config_with_overrides(test_profiles(), 8, 32768, 4, overrides);
+    let budget = Arc::clone(&config.capacity.budget);
+    let run_handle = tokio::spawn(run(config));
+
+    wait_discover_entered(&env, Duration::from_secs(2)).await;
+
+    let run_id = RunId::new_v4();
+    push_job(&env, run_id, "vm0/default", Some(minimal_context(run_id)));
+
+    // Wait for the original job to be claimed and blocked at the sandbox gate.
+    let _token = wait_cancel_token(&env.cancel_tokens, run_id, Duration::from_secs(5)).await;
+    wait_discover_entered(&env, Duration::from_secs(5)).await;
+
+    // Push the same run_id again (simulates duplicate discovery).
+    // Budget has room, but cancel_tokens already contains this run_id →
+    // the duplicate is rejected and budget is released.
+    env.handle
+        .discover_tx
+        .send(crate::provider::JobCandidate::new(
+            run_id,
+            "vm0/default".into(),
+        ))
+        .unwrap();
+    wait_discover_entered(&env, Duration::from_secs(5)).await;
+
+    // Wait for the original job to complete.
+    gate.notify_one();
+    let completion = env
+        .handle
+        .wait_completion(run_id, Duration::from_secs(5))
+        .await;
+    assert!(completion.is_some(), "original job should complete");
+
+    // Only one completion should exist for this run_id.
+    {
+        let comps = env.handle.completions.lock().unwrap();
+        let count = comps.iter().filter(|c| c.run_id == run_id).count();
+        assert_eq!(
+            count, 1,
+            "duplicate discovery should not produce a second completion"
+        );
+    }
+    wait_budget_count(&budget, 0, Duration::from_secs(5)).await;
+
+    shutdown(&env, run_handle).await;
+}
